+++ test/stress.c
@@ -27,16 +27,16 @@ void stress_small(unsigned bits, unsigned max, unsigned count) {
     bitset_offset *offsets = bitset_malloc(sizeof(bitset_offset) * bits);
 
     start = (float) clock();
-    for (unsigned j = 0; j < count; j++) {
-        for (unsigned j = 0; j < bits; j++) {
+    for (size_t j = 0; j < count; j++) {
+        for (size_t j = 0; j < bits; j++) {
             offsets[j] = bitset_rand() % max;
         }
         b = bitset_new_bits(offsets, bits);
-        for (unsigned j = 0; j < bits; j++) {
+        for (size_t j = 0; j < bits; j++) {
             offsets[j] = bitset_rand() % max;
         }
         b2 = bitset_new_bits(offsets, bits);
-        for (unsigned j = 0; j < bits; j++) {
+        for (size_t j = 0; j < bits; j++) {
             offsets[j] = bitset_rand() % max;
         }
         b3 = bitset_new_bits(offsets, bits);
@@ -56,7 +56,7 @@ void stress_small(unsigned bits, unsigned max, unsigned count) {
 
 void stress_vector(unsigned bitsets, unsigned bits, unsigned max) {
     float start, end, size = 0;
-    unsigned i;
+    size_t i;
 
     bitset_t **b = bitset_malloc(sizeof(bitset_t *) * bitsets);
     bitset_offset *offsets = bitset_malloc(sizeof(bitset_offset) * bits);
@@ -65,7 +65,7 @@ void stress_vector(unsigned bitsets, unsigned bits, unsigned max) {
     //Create the bitsets
     start = (float) clock();
     for (i = 0; i < bitsets; i++) {
-        for (unsigned j = 0; j < bits; j++) {
+        for (size_t j = 0; j < bits; j++) {
             offsets[j] = bitset_rand() % max;
         }
         b[i] = bitset_new_bits(offsets, bits);
@@ -114,8 +114,8 @@ void stress_exec(unsigned bitsets, unsigned bits, unsigned max) {
 
     //Create the bitsets
     start = (float) clock();
-    for (unsigned i = 0; i < bitsets; i++) {
-        for (unsigned j = 0; j < bits; j++) {
+    for (size_t i = 0; i < bitsets; i++) {
+        for (size_t j = 0; j < bits; j++) {
             offsets[j] = bitset_rand() % max;
         }
         b[i] = bitset_new_bits(offsets, bits);
@@ -128,7 +128,7 @@ void stress_exec(unsigned bitsets, unsigned bits, unsigned max) {
     //Estimate count based on size
     start = (float) clock();
     bitset_offset count = 0;
-    for (unsigned i = 0; i < bitsets; i++) {
+    for (size_t i = 0; i < bitsets; i++) {
         count += b[i]->length;
     }
     printf("Estimated bit count => " bitset_format "\n", count);
@@ -138,7 +138,7 @@ void stress_exec(unsigned bitsets, unsigned bits, unsigned max) {
     //Pop count all bitsets
     start = (float) clock();
     count = 0;
-    for (unsigned i = 0; i < bitsets; i++) {
+    for (size_t i = 0; i < bitsets; i++) {
         count += bitset_count(b[i]);
     }
     printf("Bit count => " bitset_format "\n", count);
@@ -148,7 +148,7 @@ void stress_exec(unsigned bitsets, unsigned bits, unsigned max) {
     //Bitwise OR + pop count
     start = (float) clock();
     bitset_operation_t *op = bitset_operation_new(b[0]);
-    for (unsigned i = 1; i < bitsets; i++) {
+    for (size_t i = 1; i < bitsets; i++) {
         bitset_operation_add(op, b[i], BITSET_OR);
     }
     printf("Unique bit count => " bitset_format "\n", bitset_operation_count(op));
@@ -162,7 +162,7 @@ void stress_exec(unsigned bitsets, unsigned bits, unsigned max) {
     //Use linear counting
     start = (float) clock();
     bitset_linear_t *e = bitset_linear_new(max);
-    for (unsigned i = 0; i < bitsets; i++) {
+    for (size_t i = 0; i < bitsets; i++) {
         bitset_linear_add(e, b[i]);
     }
     printf("Unique bit count => %u\n", bitset_linear_count(e));
@@ -170,7 +170,7 @@ void stress_exec(unsigned bitsets, unsigned bits, unsigned max) {
     printf("Executed linear count in %.2fs (%.2fMB/s)\n", end, size/end);
     bitset_linear_free(e);
 
-    for (unsigned i = 0; i < bitsets; i++) {
+    for (size_t i = 0; i < bitsets; i++) {
         bitset_free(b[i]);
     }
     bitset_malloc_free(b);
